requirements.txt
python-binance==0.6.2
slackclient==1.1.0
